+++ Minicon.SevDesk.Client/Api/IReportApi.cs
@@ -18,6 +18,7 @@ public interface IReportApi
 	/// <param name="sevQuery"></param>
 	/// <param name="download"> (optional)</param>
 	/// <returns>Task of Object</returns>
+	[Get("/Report/contactlist")]
 	Task<object> ReportContactAsync(ReportContactRequest sevQuery, bool? download = null);
 
 	/// <summary>